# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type InterviewSessionExperienceLevel =
  (typeof InterviewSessionExperienceLevel)[keyof typeof InterviewSessionExperienceLevel];

export const InterviewSessionExperienceLevel = {
  entry: "entry",
  mid: "mid",
  senior: "senior",
  lead: "lead",
} as const;

export type InterviewSessionDifficulty =
  (typeof InterviewSessionDifficulty)[keyof typeof InterviewSessionDifficulty];

export const InterviewSessionDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export type InterviewSessionStatus =
  (typeof InterviewSessionStatus)[keyof typeof InterviewSessionStatus];

export const InterviewSessionStatus = {
  generating: "generating",
  active: "active",
  completed: "completed",
  failed: "failed",
} as const;

export interface InterviewSession {
  id: string;
  userId: string;
  jobRole: string;
  companyName?: string | null;
  experienceLevel: InterviewSessionExperienceLevel;
  interviewRounds: number;
  roundDetails?: string | null;
  difficulty: InterviewSessionDifficulty;
  status: InterviewSessionStatus;
  questionCount: number;
  startedAt: string;
  endedAt?: string | null;
  score?: number | null;
}

export type QuestionDifficulty =
  (typeof QuestionDifficulty)[keyof typeof QuestionDifficulty];

export const QuestionDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface Question {
  id: string;
  content: string;
  category: string;
  roundName: string;
  sequence: number;
  difficulty: QuestionDifficulty;
}

export interface Answer {
  id: string;
  sessionId: string;
  questionId: string;
  question?: Question | null;
  /**
   * @minLength 1
   * @maxLength 12000
   */
  response: string;
  createdAt: string;
}

export interface CategoryScore {
  category: string;
  /**
   * @minimum 0
   * @maximum 100
   */
  score: number;
}

export interface DimensionScores {
  /**
   * @minimum 0
   * @maximum 100
   */
  relevance: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  accuracy: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  depth: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  structure: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  communication: number;
}

export interface AnswerScore {
  questionId: string;
  /**
   * @minimum 0
   * @maximum 100
   */
  score: number;
  evidence: string[];
  missedSignals: string[];
  improvedAnswer: string;
  dimensionScores: DimensionScores;
}

export interface Feedback {
  id: string;
  sessionId: string;
  /**
   * @minimum 0
   * @maximum 100
   */
  score: number;
  strengths: string[];
  weaknesses: string[];
  suggestions: string[];
  categoryScores: CategoryScore[];
  answerBreakdown: AnswerScore[];
  methodology: string;
  summary: string;
  createdAt: string;
}

export type InterviewSessionDetail = InterviewSession & {
  answers?: Answer[];
  feedback?: Feedback | null;
};

export type CreateInterviewBodyDifficulty =
  (typeof CreateInterviewBodyDifficulty)[keyof typeof CreateInterviewBodyDifficulty];

export const CreateInterviewBodyDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export type CreateInterviewBodyExperienceLevel =
  (typeof CreateInterviewBodyExperienceLevel)[keyof typeof CreateInterviewBodyExperienceLevel];

export const CreateInterviewBodyExperienceLevel = {
  entry: "entry",
  mid: "mid",
  senior: "senior",
  lead: "lead",
} as const;

export interface DocumentUpload {
  /** @maxLength 255 */
  filename: string;
  mimeType: string;
  base64: string;
}

export interface CreateInterviewBody {
  /**
   * @minLength 2
   * @maxLength 120
   */
  jobRole: string;
  /** @maxLength 120 */
  companyName?: string;
  difficulty: CreateInterviewBodyDifficulty;
  experienceLevel: CreateInterviewBodyExperienceLevel;
  /**
   * @minimum 1
   * @maximum 10
   */
  questionCount?: number;
  /**
   * @minimum 1
   * @maximum 8
   */
  interviewRounds: number;
  /** @maxLength 5000 */
  roundDetails?: string;
  /** @maxLength 30000 */
  jobDescription?: string;
  jobDescriptionFile?: DocumentUpload;
  resumeFile?: DocumentUpload;
}

export type PlanRequestBody = CreateInterviewBody;

export interface SubmitAnswerBody {
  questionId: string;
  response: string;
}

export interface TranscribeAnswerBody {
  /** @maxLength 14000000 */
  audioBase64: string;
  mimeType: string;
}

export interface TranscriptionResult {
  transcript: string;
}

export interface PlanRound {
  name: string;
  purpose: string;
  focusAreas: string[];
  estimatedDays: number;
}

export interface PlanDay {
  day: number;
  title: string;
  tasks: string[];
}

export interface PlanSource {
  title: string;
  url: string;
}

export interface InterviewPlan {
  id: string;
  companyName?: string | null;
  jobRole: string;
  experienceLevel: string;
  interviewRounds: number;
  candidateSummary: string;
  roleSummary: string;
  researchSummary: string;
  rounds: PlanRound[];
  schedule: PlanDay[];
  sources: PlanSource[];
  createdAt: string;
}

export type AnalyticsOverviewScoresByDifficulty = {
  easy?: number | null;
  medium?: number | null;
  hard?: number | null;
};

export type HistoryItemDifficulty =
  (typeof HistoryItemDifficulty)[keyof typeof HistoryItemDifficulty];

export const HistoryItemDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export type HistoryItemStatus =
  (typeof HistoryItemStatus)[keyof typeof HistoryItemStatus];

export const HistoryItemStatus = {
  generating: "generating",
  active: "active",
  completed: "completed",
  failed: "failed",
} as const;

export interface HistoryItem {
  id: string;
  jobRole: string;
  difficulty: HistoryItemDifficulty;
  score?: number | null;
  questionCount: number;
  startedAt: string;
  endedAt?: string | null;
  status: HistoryItemStatus;
}

export interface AnalyticsOverview {
  totalSessions: number;
  completedSessions: number;
  averageScore?: number | null;
  bestScore?: number | null;
  topCategory?: string | null;
  recentActivity: HistoryItem[];
  scoresByDifficulty: AnalyticsOverviewScoresByDifficulty;
}