# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /interviews:
    get:
      operationId: listInterviews
      tags: [interviews]
      summary: List user interview sessions
      responses:
        "200":
          description: List of interview sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InterviewSession"
    post:
      operationId: createInterview
      tags: [interviews]
      summary: Start a new interview session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInterviewBody"
      responses:
        "201":
          description: Created interview session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewSession"
  /interviews/{id}:
    get:
      operationId: getInterview
      tags: [interviews]
      summary: Get a specific interview session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Interview session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewSessionDetail"
    patch:
      operationId: endInterview
      tags: [interviews]
      summary: End an interview session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated interview session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewSession"
  /interviews/{id}/next-question:
    get:
      operationId: getNextQuestion
      tags: [questions]
      summary: Get next question for a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Next question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "204":
          description: No more questions available
  /interviews/{id}/answers:
    post:
      operationId: submitAnswer
      tags: [answers]
      summary: Submit an answer for a question in a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAnswerBody"
      responses:
        "201":
          description: Saved answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Answer"
  /interviews/{id}/feedback:
    get:
      operationId: getFeedback
      tags: [feedback]
      summary: Get feedback for an interview session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Feedback for session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"
    post:
      operationId: generateFeedback
      tags: [feedback]
      summary: Generate AI feedback for a completed interview session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Generated feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"
  /interviews/{id}/transcribe:
    post:
      operationId: transcribeAnswer
      tags: [answers]
      summary: Transcribe an audio answer using AI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscribeAnswerBody"
      responses:
        "200":
          description: Transcription result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionResult"
...
components:
  schemas:
...
    InterviewSession:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        jobRole:
          type: string
        companyName:
          type: string
          nullable: true
        experienceLevel:
          type: string
          enum: [entry, mid, senior, lead]
        interviewRounds:
          type: integer
        roundDetails:
          type: string
          nullable: true
        difficulty:
          type: string
          enum: [easy, medium, hard]
        status:
          type: string
          enum: [generating, active, completed, failed]
        questionCount:
          type: integer
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
        score:
          type: number
          nullable: true
      required:
        - id
        - userId
        - jobRole
        - experienceLevel
        - interviewRounds
        - difficulty
        - status
        - questionCount
        - startedAt
    InterviewSessionDetail:
      allOf:
        - $ref: "#/components/schemas/InterviewSession"
        - type: object
          properties:
            answers:
              type: array
              items:
                $ref: "#/components/schemas/Answer"
            feedback:
              $ref: "#/components/schemas/Feedback"
              nullable: true
    CreateInterviewBody:
      type: object
      properties:
        jobRole:
          type: string
          minLength: 2
          maxLength: 120
        companyName:
          type: string
          maxLength: 120
        difficulty:
          type: string
          enum: [easy, medium, hard]
        experienceLevel:
          type: string
          enum: [entry, mid, senior, lead]
        questionCount:
          type: integer
          minimum: 1
          maximum: 10
        interviewRounds:
          type: integer
          minimum: 1
          maximum: 8
        roundDetails:
          type: string
          maxLength: 5000
        jobDescription:
          type: string
          maxLength: 30000
        jobDescriptionFile:
          $ref: "#/components/schemas/DocumentUpload"
        resumeFile:
          $ref: "#/components/schemas/DocumentUpload"
      required:
        - jobRole
        - difficulty
        - experienceLevel
        - interviewRounds
...
    DocumentUpload:
      type: object
      properties:
        filename:
          type: string
          maxLength: 255
        mimeType:
          type: string
        base64:
          type: string
      required: [filename, mimeType, base64]
    Question:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        category:
          type: string
        roundName:
          type: string
        sequence:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium, hard]
      required:
        - id
        - content
        - category
        - roundName
        - sequence
        - difficulty
    Answer:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        questionId:
          type: string
        question:
          $ref: "#/components/schemas/Question"
          nullable: true
        response:
          type: string
          minLength: 1
          maxLength: 12000
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - questionId
        - response
        - createdAt
    SubmitAnswerBody:
      type: object
      properties:
        questionId:
          type: string
        response:
          type: string
      required:
        - questionId
        - response
    TranscribeAnswerBody:
      type: object
      properties:
        audioBase64:
          type: string
          maxLength: 14000000
        mimeType:
          type: string
      required:
        - audioBase64
        - mimeType
    TranscriptionResult:
      type: object
      properties:
        transcript:
          type: string
      required:
        - transcript
    Feedback:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        score:
          type: number
          minimum: 0
          maximum: 100
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        suggestions:
          type: array
          items:
            type: string
        categoryScores:
          type: array
          items:
            $ref: "#/components/schemas/CategoryScore"
        answerBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/AnswerScore"
        methodology:
          type: string
        summary:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - score
        - strengths
        - weaknesses
        - suggestions
        - categoryScores
        - answerBreakdown
        - methodology
        - summary
        - createdAt
    CategoryScore:
      type: object
      properties:
        category:
          type: string
        score:
          type: number
          minimum: 0
          maximum: 100
      required: [category, score]
    AnswerScore:
      type: object
      properties:
        questionId:
          type: string
        score:
          type: number
          minimum: 0
          maximum: 100
        evidence:
          type: array
          items:
            type: string
        missedSignals:
          type: array
          items:
            type: string
        improvedAnswer:
          type: string
        dimensionScores:
          $ref: "#/components/schemas/DimensionScores"
      required:
        - questionId
        - score
        - evidence
        - missedSignals
        - improvedAnswer
        - dimensionScores
    DimensionScores:
      type: object
      properties:
        relevance: { type: number, minimum: 0, maximum: 100 }
        accuracy: { type: number, minimum: 0, maximum: 100 }
        depth: { type: number, minimum: 0, maximum: 100 }
        structure: { type: number, minimum: 0, maximum: 100 }
        communication: { type: number, minimum: 0, maximum: 100 }
      required: [relevance, accuracy, depth, structure, communication]
...